// src/types/carrier/transport.ts
export interface VehicleAVBResource {
  id: number;
  day: string; //yyyy-mm-dd
  price: number;
  status: number;
  giai: string;
  trsp_cli_prty_name_txt: string;
  trans_type: number;
  operator_id: string;
  operator_name: string;
  operator_code: string;
  car_info_id: number;
  vehicle_availability_resource_id: number;
  departure_from: number;
  arrival_to: number;
  trip_name: string;
  departure_time_min: string;
  departure_time_max: string;
  arrival_time: string;
  service_strt_time: string;
  service_end_time: string;
  adjustment_price: number;
  vehicle_type: number | null;
  display_order: number | null;
  assign_type: number;
  max_payload: number;
  total_width: number;
  total_height: number;
  total_length: number;
  vehicle_size: number;
  vehicle_name: string;
  cut_off_fee: number;
  cut_off_info_id: number;
  cut_off_time: number;
  free_time_fee: number;
  free_time_info_id: number;
  free_time_time: number;
  temperature_range: string[] | null;
  tractor_idcr: number;
  car_license_plt_num_id: string;
  trailer_license_plt_num_id: string;
  propose_trsp_plan: ProposeTrspPlan[] | null;
}
// ...
export interface CarInfo {
  id: number;
  giai: string;
  service_no: string;
  pcke_frm_cd: null;
  car_wid_meas: number;
  freight_rate: number;
  service_name: string;
  tractor_idcr: null;
  car_hght_meas: number;
  car_lngh_meas: number;
  car_weig_meas: null;
  car_body_num_cd: null;
  car_ctrl_num_id: null;
  pcke_frm_name_cd: null;
  service_end_date: null;
  service_end_time: null;
  car_cls_of_shp_cd: null;
  service_strt_date: string;
  service_strt_time: string;
  car_cls_of_size_cd: null;
  car_max_untl_cp_quan: null;
  car_trk_bed_wid_meas: null;
  trms_of_lwr_tmp_meas: null;
  trms_of_upp_tmp_meas: null;
  car_max_load_vol_meas: null;
  car_trk_bed_hght_meas: null;
  car_trk_bed_lngh_meas: null;
  car_cls_of_crn_exst_cd: null;
  car_cls_of_rfg_exst_cd: null;
  car_license_plt_num_id: null;
  car_rmk_about_eqpm_txt: null;
  car_vol_of_hzd_item_meas: number | null;
  car_trk_bed_grnd_hght_meas: number | null;
  trailer_license_plt_num_id: string | null;
  car_cls_of_tlg_lftr_exst_cd: string | null;
  car_max_load_capacity1_meas: number;
  car_max_load_capacity2_meas: number | null;
  car_cls_of_wing_body_exst_cd: number | null;
  car_spc_grv_of_hzd_item_meas: number | null;
  car_cmpn_name_of_gtp_crtf_exst_txt: string | null;
  vehicle_name: string | null;
  temperature_range: string[] | null;
  display_order: number | null;
  cut_off_fee: number;
  cut_off_time: number;
  order_status: number;
}

export interface ProposeTrspPlan {
  id: number;
  arrival_to: number;
  cns_line_item_id: number;
  collection_date: string;
  collection_time_from: string;
  collection_time_to: string;
  outer_package_code: number;
  price_per_unit: number;
  status: number;
  departure_from: number;
  operator_id: string;
  transport_code: string;
  transport_name: string;
  special_instructions: string;
  vehicle_condition: string;
  total_height: number;
  total_length: number;
  total_width: number;
  trailer_number: number;
  trailer_number_rest: number;
  weight: number;
  tractor_idcr: number;
  car_ctrl_num_id: number;
  car_license_plt_num_id: number;
  trailer_license_plt_num_id: number;
  created_date: string;
  service_strt_date: string;
  service_strt_time: string;
  service_end_time: string;
  operator_name: string;
  trsp_plan_line_item: TrspPlanLineItem[] | null;
  shipper_info?: IShipperInfo;
}

export interface TrspPlanLineItem {
  id: number;
  aped_arr_from_date: string;
  aped_arr_from_time_prfm_dttm: string;
  aped_arr_time_trms_srvc_rqrm_cd: string;
  aped_arr_to_date: string;
  aped_arr_to_time_prfm_dttm: string;
  car_cls_of_crn_exst_cd: string;
  car_cls_of_rfg_exst_cd: string;
  car_cls_of_shp_cd: string;
  car_cls_of_size_cd: string;
  car_cls_of_tlg_lftr_exst_cd: string;
  car_cls_of_wing_body_exst_cd: string;
  car_cls_prio_cd: string;
  car_rmk_about_eqpm_txt: string;
  cls_of_carg_in_srvc_rqrm_cd: string;
  cls_of_pkg_up_srvc_rqrm_cd: string;
  cmn_inv_num_id: string;
  cnee_prim_cnt_pers_name_txt: string;
  cnee_prty_brnc_off_id: string;
  cnee_prty_head_off_id: string;
  cnee_prty_name_txt: string;
  cnee_pstc_cd: string;
  cnee_pstl_adrs_line_one_txt: string;
  cnee_sct_id: string;
  cnee_sct_name_txt: string;
  cnee_tel_cmm_cmp_num_txt: string;
  cnsg_prty_brnc_off_id: string;
  cnsg_prty_head_off_id: string;
  cnsg_prty_name_txt: string;
  cnsg_pstc_cd: string;
  cnsg_pstl_adrs_line_one_txt: string;
  cnsg_sct_sped_org_id: string;
  cnsg_sct_sped_org_name_txt: string;
  cnsg_tel_cmm_cmp_num_txt: string;
  del_note_id: string;
  dsed_cll_from_date: string;
  dsed_cll_from_time: string | null;
  dsed_cll_time_trms_srvc_rqrm_cd: string;
  dsed_cll_to_date: string;
  dsed_cll_to_time: string | null;
  freight_rate: number;
  fret_clim_to_prty_brnc_off_id: string;
  fret_clim_to_prty_head_off_id: string;
  fret_clim_to_prty_name_txt: string;
  fret_clim_to_sct_sped_org_id: string;
  fret_clim_to_sct_sped_org_name_txt: string;
  inv_num_id: string;
  istd_totl_pcks_quan: number;
  istd_totl_untl_quan: number;
  istd_totl_vol_meas: number;
  istd_totl_weig_meas: number;
  logs_srvc_prv_prty_brnc_off_id: string;
  logs_srvc_prv_prty_head_off_id: string;
  logs_srvc_prv_prty_name_txt: string;
  logs_srvc_prv_sct_prim_cnt_pers_name_txt: string;
  logs_srvc_prv_sct_sped_org_id: string;
  logs_srvc_prv_sct_sped_org_name_txt: string;
  logs_srvc_prv_sct_tel_cmm_cmp_num_txt: string;
  mix_load_num_id: string;
  num_unt_cd: string;
  operator_id: string;
  pyr_cls_srvc_rqrm_cd: string;
  rced_ord_num_id: string;
  road_carr_arr_sped_org_id: string;
  road_carr_arr_sped_org_name_txt: string;
  road_carr_depa_sped_org_id: string;
  road_carr_depa_sped_org_name_txt: string;
  road_carr_srvc_typ_cd: string;
  service_end_date: string;
  service_end_time: string;
  service_name: string;
  service_no: string;
  service_strt_date: string;
  service_strt_time: string;
  shpm_num_id: string;
  trms_of_lwr_tmp_meas: number;
  trms_of_mix_load_cnd_cd: string;
  trms_of_mix_load_txt: string;
  trms_of_upp_tmp_meas: number;
  trsp_cli_prty_brnc_off_id: string;
  trsp_cli_prty_head_off_id: string;
  trsp_cli_prty_name_txt: string;
  trsp_cli_tel_cmm_cmp_num_txt: string;
  trsp_instruction_date_subm_dttm: string;
  trsp_instruction_id: string;
  trsp_means_typ_cd: string;
  trsp_plan_msg_info_id: number;
  trsp_root_prio_cd: string;
  trsp_rqr_prty_brnc_off_id: string;
  trsp_rqr_prty_head_off_id: string;
  trsp_rqr_prty_name_txt: string;
  trsp_rqr_pstc_cd: string;
  trsp_rqr_pstl_adrs_line_one_txt: string;
  trsp_rqr_sct_sped_org_id: string;
  trsp_rqr_sct_sped_org_name_txt: string;
  trsp_rqr_sct_tel_cmm_cmp_num_txt: string;
  trsp_srvc_note_one_txt: string;
  trsp_srvc_note_two_txt: string;
  trsp_srvc_typ_cd: string;
  vol_unt_cd: string;
  weig_unt_cd: string;
  tractor_idcr?: number;
  car_license_plt_num_id?: number;
  trailer_license_plt_num_id?: number;
  vehicle_name?: string | null;
  temperature_range?: string[] | null;
  display_order?: number | null;
  is_matched: boolean;
  car_info: CarInfo[] | null;
}

export interface IShipperInfo {
  id: number;
  operator_name: string;
  outer_package_code: string;
  collection_date: string;
  collection_time_from: string;
  collection_time_to: string;
  special_instructions: string;
  temperature_range: string[];
}